[REFACTOR] ClothPickerView 의 subView 들을 다른 파일로 이동
