+++ POMPOM/POMPOM/UI/ClothPickerView/ClothGrid.swift
@@ -1,0 +1,56 @@
+//
+//  ClothGrid.swift
+//  POMPOM
+//
+//  Created by GOngTAE on 2022/06/15.
+//
+
+import SwiftUI
+
+struct ClothGrid: View {
+    @ObservedObject var vm: PickerViewModel
+    @Binding var currentCategory: ClothCategory
+    @Binding var currentHex: String
+    
+    let columns = [
+        GridItem(.flexible(minimum: 60), spacing: 20),
+        GridItem(.flexible(minimum: 60), spacing: 20)
+    ]
+    
+    var body: some View {
+        LazyVGrid(columns: columns, spacing: 20) {
+            ForEach($vm.currentItems, id: \.self) { item in
+                ZStack {
+                    Image(vm.imageName(name: item.wrappedValue) + "B")
+                        .resizable()
+                        .aspectRatio(1, contentMode: .fit)
+                        .foregroundColor(Color(hex: currentHex))
+                    
+                    Image(vm.imageName(name: item.wrappedValue))
+                        .resizable()
+                        .aspectRatio(1, contentMode: .fit)
+                        .foregroundColor(currentHex == "000000" ? .gray : .black) // 검정색일 때 옷 테두리 색상 변경
+                        .overlay {
+                            if vm.selectedItems[currentCategory]?.id == item.wrappedValue {
+                                RoundedRectangle(cornerRadius: 20)
+                                    .stroke(Color(hex: "BABABA"), lineWidth: 4)
+                                    .frame(width: 180, height: 180, alignment: .center)
+                            }
+                        }
+                }
+                .onTapGesture {
+                    withAnimation {
+                        vm.selectItem(name: item.wrappedValue, hex: currentHex)
+                    }
+                }
+            }
+        }
+    }
+}
+
+
+struct ClothGrid_Previews: PreviewProvider {
+    static var previews: some View {
+        ClothGrid()
+    }
+}
