+++ POMPOM/POMPOM/UI/ClothPickerView/ColorGrid.swift
@@ -1,0 +1,58 @@
+//
+//  ColorGrid.swift
+//  POMPOM
+//
+//  Created by GOngTAE on 2022/06/15.
+//
+
+import SwiftUI
+
+struct ColorGrid: View {
+    @ObservedObject var vm: PickerViewModel
+    @Binding var currentHex: String
+    
+    let rows = [GridItem(.fixed(44), spacing: 20)]
+    var body: some View {
+        LazyHGrid(rows: rows, spacing: 18) {
+            ForEach($vm.currentPresets, id: \.self) { item in
+                ZStack {
+                    Circle()
+                        .fill(Color(hex: item.wrappedValue))
+                        .frame(width: 44)
+                        .shadow(radius: 5)
+                        .onTapGesture {
+                            withAnimation {
+                                currentHex = item.wrappedValue
+                            }
+                        }
+                        .overlay {
+                            if item.wrappedValue == currentHex {
+                                Circle()
+                                    .stroke(Color(hex: "BABABA"), lineWidth: 3)
+                                    .frame(width: 60, height: 60, alignment: .center)
+                            }
+                        }
+                }
+            }
+            Circle()
+                .fill(Color(hex: "D8D8D8"))
+                .frame(width: 44)
+                .overlay(
+                    Image(systemName: "plus")
+                        .font(.title2)
+                        
+                )
+                .onTapGesture {
+                    print("DEBUG : Button tapped")
+                    vm.addPreset(hex: "23F323")
+                }
+        }
+    }
+}
+
+
+struct ColorGrid_Previews: PreviewProvider {
+    static var previews: some View {
+        ColorGrid()
+    }
+}
